# server.js/src/swagger/v3/swaggerV3.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Reprography System",
    "description": "PT: Sistema para solicitação e controle de reprografias. \n EN: System for requesting and controlling reprographs.",
    "contact": {}
  },
  "host": "localhost:3002",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/myRequests/rated={rated}": {
      "get": {
        "summary": "Meus Pedidos",
        "tags": [
          "USER"
        ],
        "operationId": "MeusPedidos",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "rated",
	    "in": "path",
            "required": true,
            "type": "string",
            "description": "0 = ainda não avaliados, 1 = avaliados"
          },
          {
            "name": "accessToken",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Resgate o token de acesso na rota de login"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/Nenhumpedidoencontrado1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Nenhum pedido encontrado!"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Access-Control-Allow-Headers": {
                "type": "string",
                "default": "accessToken, Origin, Content-Type, Accept"
              },
              "Content-Length": {
                "type": "string",
                "default": "39"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"27-PJcGEVeLG7MDvv5/roSsMNDCgrQ\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 03 Nov 2021 12:26:46 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "default": "",
              "items": {
                "$ref": "#/definitions/Meuspedidosavaliado"
              }
            },
            "examples": {
              "application/json; charset=utf-8": [
                {
                  "id_pedido": 2,
                  "titulo_pedido": "Livro XYZ",
                  "nif": 123,
                  "id_modo_envio": 1,
                  "id_avaliacao_pedido": 1,
                  "avaliacao_obs": "Atendeu muito bem!",
                  "custo_total": 37.5,
                  "createdAt": "2021-11-03 09:42:26",
                  "updatedAt": "2021-11-03 09:43:05"
                },
                {
                  "id_pedido": 4,
                  "titulo_pedido": "Livro XYZ",
                  "nif": 123,
                  "id_modo_envio": 1,
                  "id_avaliacao_pedido": 2,
                  "avaliacao_obs": "Não me atendeu...",
                  "custo_total": 10.48,
                  "createdAt": "2021-11-03 09:44:18",
                  "updatedAt": "2021-11-03 09:44:29"
                }
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Access-Control-Allow-Headers": {
                "type": "string",
                "default": "accessToken, Origin, Content-Type, Accept"
              },
              "Content-Length": {
                "type": "string",
                "default": "442"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1ba-rzOZshH4tk9FlNfvlyXoUkBuaUU\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 03 Nov 2021 12:45:13 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "schema": {
              "$ref": "#/definitions/Par%C3%A2metropararatedinv%C3%A1lido1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Insira um parâmetro válido!"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Access-Control-Allow-Headers": {
                "type": "string",
                "default": "accessToken, Origin, Content-Type, Accept"
              },
              "Content-Length": {
                "type": "string",
                "default": "43"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2b-MrZ1PY0F3S6H/FM8SfnDrkeHDC8\""
              },
              "Date": {
                "type": "string",
                "default": "Wed, 03 Nov 2021 20:42:49 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Nenhumpedidoencontrado1": {
      "title": "Nenhumpedidoencontrado1",
      "example": {
        "message": "Nenhum pedido encontrado!"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },
    "Meuspedidosavaliado": {
      "title": "Meuspedidosavaliado",
      "example": {
        "id_pedido": 2,
        "titulo_pedido": "Livro XYZ",
        "nif": 123,
        "id_modo_envio": 1,
        "id_avaliacao_pedido": 1,
        "avaliacao_obs": "Atendeu muito bem!",
        "custo_total": 37.5,
        "createdAt": "2021-11-03 09:42:26",
        "updatedAt": "2021-11-03 09:43:05"
      },
      "type": "object",
      "properties": {
        "id_pedido": {
          "type": "integer",
          "format": "int32"
        },
        "titulo_pedido": {
          "type": "string"
        },
        "nif": {
          "type": "integer",
          "format": "int32"
        },
        "id_modo_envio": {
          "type": "integer",
          "format": "int32"
        },
        "id_avaliacao_pedido": {
          "type": "integer",
          "format": "int32"
        },
        "avaliacao_obs": {
          "type": "string"
        },
        "custo_total": {
          "type": "number",
          "format": "double"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "required": [
        "id_pedido",
        "titulo_pedido",
        "nif",
        "id_modo_envio",
        "id_avaliacao_pedido",
        "avaliacao_obs",
        "custo_total",
        "createdAt",
        "updatedAt"
      ]
    },
...
  },
...
}